Simplify AST expansion.
In the final stage of expanding an iteration with a finite upper bound,
use the equivalent of (.(.(.)?)?)? instead of (|.(|.(|.))).  This saves
one AST node per iteration, reducing memory usage and very slightly
speeding up both AST expansion and AST-to-TNFA translation, although
the resulting TNFA appears unchanged in the cases I examined.  In
practical terms, retest and wretest both run a little over 5% faster
with this patch than without it.