Compute nullability up-front.
All the information needed to decide whether an AST node is nullable is
available at creation, so do it right away instead of delaying until
`tre_compute_nfl()`.  Change the latter to assert nullability instead of
setting it; we can change that later, when we have more confidence in the
change.