Recompute node positions after expansion.
During AST expansion, it is possible for new literals to be created and
given the same position as existing literals.  The code attempts to
reposition nodes when this happens but does not always succeed.  It is
also possible for positions to be left unused.  The quick fix is to do a
single pass over the tree after expansion and renumber all literals
sequentially.